requirements.txt
pytest>=7.4.0
pytest-cov>=4.1.0
alembic>=1.12.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
pre-commit>=3.5.0
black>=23.3.0
flake8>=6.0.0
pytest-asyncio>=0.21.1
pytest-xdist>=3.3.1
pytest-timeout>=2.1.0
pytest-env>=1.1.1
